requirements.txt
python-dotenv
google-generativeai
textblob
vaderSentiment
regex
requests
numpy
flask